+++ packages/moonstone/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The following is a curated list of changes in the Enact moonstone module, newest changes on the top.
 
+## [unrelease]
+
+### Fixed
+
+- `moonstone/Notification` to show all buttons in one line
+
 ## [3.2.3] - 2019-11-01
 
 ### Changed