+++ packages/moonstone/CHANGELOG.md
@@ -7,6 +7,7 @@ The following is a curated list of changes in the Enact moonstone module, newest
 ### Fixed
 
 - `moonstone/Notification` line height for non-latin locales
+- `moonstone/Notification` to show all buttons in one line
 
 ## [3.2.3] - 2019-11-01